==> code/inference_fb_2022/02_fb2022_inf_binary.ipynb <==
# this model has a problem, it classifies most ads as one category

import sklearn.model_selection as ms
from sklearn.pipeline import Pipeline
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.calibration import CalibratedClassifierCV
from sklearn.naive_bayes import MultinomialNB
from sklearn import metrics
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.model_selection import train_test_split
import pandas as pd
import numpy as np
from joblib import dump, load
from tqdm import tqdm

dir_models = '../../models/binary_rf_v1'
path_inference_set = '../../data/fb_22_for_inf.csv'
path_inference_set_output = '../../data/fb_22_output_binary.csv'

#----
# Load the inference dataset
text_field = 'text'
df = pd.read_csv(path_inference_set)
df = df.dropna(subset = [text_field]) # remove NAs
df = df[df[text_field] != '_error'] # remove errors
df = df.reset_index(drop = True)

# Load the variable labels
with open('../../data/issue_labels_65.txt', 'r') as reader:
  labels = reader.read().split('\n')
labels = labels[:-1]


preds_l = []

# Batch size for predictions
batch_size = 512

# Create a single progress bar for both loops
total_labels = len(labels)
total_rows = len(df[text_field])

with tqdm(total=total_labels * total_rows, desc='Making Predictions') as pbar:
    for l in labels:
        clf_rf = load(dir_models + '/issues_rf_' + l + '.joblib')

        for i in range(0, total_rows, batch_size):
            # Batch predict using NumPy for efficiency
            batch_transcripts = df[text_field][i:i + batch_size].values
            batch_predictions = clf_rf.predict(batch_transcripts)
            
            preds_l.extend(batch_predictions)
            pbar.update(len(batch_transcripts))  # Update the single progress bar for each batch

# Ensure the progress bar reaches 100% at the end
pbar.n = total_labels * total_rows
pbar.last_print_n = total_labels * total_rows
pbar.refresh()

df_preds = pd.DataFrame({y: x for x, y in zip(preds_l, labels)}, index=df.index)
df_results = pd.concat([df, df_preds], axis=1)

df_results.to_csv(path_inference_set_output)

df_results
