==> code/inference_fb_2022/02_fb2022_inf_multilabel.ipynb <==
# Inference dataset (change)
inference_dataset = '../../data/fb_22_for_inf.csv'
text_field = 'text'
output_file = '../../data/fb_22_output_multi.csv.gz'

# Variable label file (don't change)
label_file = '../../data/issue_labels_25.txt'

# Model files (usually don't change)
model_pytorch_model = '../../models/multilabel_trf_v4/model.safetensors'
model_config = '../../models/multilabel_trf_v4/config.json'

#!pip install pandas==2.2.2
#!pip install scikit-learn==1.0.2
#!pip install numpy==1.26.4
#!pip install joblib==1.4.2
#!pip install torch==2.3.1
#!pip install tqdm==4.66.4
#!pip install transformers==4.41.2
#!pip install datasets==2.20.0
#!pip install ipywidgets

from tqdm import tqdm
import os
import shutil

import numpy as np
import pandas as pd
import torch
import transformers

from transformers import (
    AutoModelForSequenceClassification,
    AutoTokenizer,
)

#----
# Copy the model files to the model dir
shutil.copyfile(model_pytorch_model, '../../models/model.safetensors')
shutil.copyfile(model_config, '../../models/config.json')

# Copy inference data (here test set)
shutil.copyfile(inference_dataset, './inference_dataset.csv')


# Get cpu, gpu or mps device for training.
device = (
    "cuda"
    if torch.cuda.is_available()
    else "mps"
    if torch.backends.mps.is_available()
    else "cpu"
)
print(f"Using {device} device")

#----
# Load the trained model and tokenizer
model = AutoModelForSequenceClassification.from_pretrained('models').to(device)
tokenizer = AutoTokenizer.from_pretrained("distilbert-base-uncased")

#----
# Load the inference dataset
df = pd.read_csv('./inference_dataset.csv')
df = df.dropna(subset = [text_field]) # remove NAs
df = df[df[text_field] != '_error'] # remove errors
df = df.reset_index(drop = True)


# Load the variable labels
with open('../../data/issue_labels_25.txt', 'r') as reader:
    labels = reader.read().split('\n')
# They were created like this:
# df = pd.read_csv('data/issues_tv_fb_18_20.csv')
# with open('data/issue_labels_65.txt', 'w') as writer:
#   for i in df.columns[2:].tolist():
#     writer.write(i + '\n')

df.head()

#----
# Inference

# Batch the text Series (batch size 16)
texts = df[text_field].to_list()
batch_size = 16
list_df = [texts[i:i+batch_size] for i in range(0,len(texts),batch_size)]

# Use the tokenizer to encode the Series in batches
batched_examples = []
for text_chunk in tqdm(list_df, desc="Tokenizing batches", unit="batch"):
    batched_examples.append(tokenizer.batch_encode_plus(
        text_chunk, truncation=True, padding=True, return_tensors="pt"
    ))


# Batch inference
# For inference, calculating the gradients is unnecessary
# with torch.no_grad(): turns them off, which is faster (seems about 10x faster on CPU, and 2x faster on GPU or so)
outputs_list = []

for encoded_chunk in tqdm(batched_examples):
    encoded_chunk = {key: value.to(device) for key, value in encoded_chunk.items()}
    
    with torch.no_grad():
        outputs_list.append(model(**encoded_chunk))

# Convert to 1s and 0s
preds_l = []
for output in outputs_list:
    preds = output.logits.sigmoid().cpu().numpy() > 0.5
    preds = preds.astype(int)
    preds_l.append(preds)

outputs = np.vstack(preds_l)

# Convert to pd DataFrame and save
df_preds = pd.DataFrame(outputs)
df_preds.columns = labels[:]
df_results = pd.concat([df, df_preds], axis = 1)
df_results2 = df_results.dropna()
df_results3 = df_results2.drop_duplicates()

df_results3

df_results3.to_csv(output_file, index=False, compression='gzip', encoding='utf-8')
