==> code/inference_fb_2022/03_fb2022_post_process.ipynb <==
import pandas as pd
import numpy as np
import os

issue = pd.read_csv("../../data/fb_22_output_multi.csv.gz")

issue.head(2)

# Function to get column names where value is 1 for the last 25 columns
def get_active_features(row):
    active_features = issue.columns[-25:][row.iloc[-25:] == 1]
    return ','.join(active_features)

# Apply the function to each row
issue['issue'] = issue.apply(get_active_features, axis=1)

# Split ids
issue['id'] = issue['id'].str.split('|')
# "Un-deduplicate", or "Re-hydrate", in WMP lingo
issue = issue.explode('id')
# Split into ad id and field
issue_ids = issue['id'].str.split('__', expand = True)
issue_ids.columns = ['ad_id', 'field']
issue = pd.concat([issue, issue_ids], axis = 1)
issue = issue.drop(labels = ['id'], axis = 1)
issue = issue[['ad_id', 'field', 'issue']]

issue.rename(columns={'field': 'issue_field', 'issue': 'issue_class'}, inplace=True)

# Replace various representations of missing values with NaN
missing_values = ['', None, 'NA', 'missing']
issue['issue_class'] = issue['issue_class'].replace(missing_values, np.nan)
issue = issue.dropna(subset=['issue_class'])

agg_issue = issue.groupby('ad_id').agg({'issue_field': lambda x: '|'.join(x),
                                       'issue_class': lambda x: '|'.join(x)}).reset_index()

agg_issue.head(5)

agg_issue.shape

agg_issue.to_csv("../../data/fb2022_output_processed.csv", index = False)
